+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Ability to Enable/Disable many objects at once
+
 ### Fixed
 
 - Fixed changes to Catalogue visibility checkboxes not being persisted